pyproject.toml
[tool.poetry]
name = "vpn-tester"
version = "0.1.0"
description = ""
authors = ["Ian <exor2008@yandex.ru>"]
readme = "README.md"
packages = [{include = "vpn_tester"}]

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "^44.0.0"
scapy = "^2.6.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
